fix: resolve audit findings and complete test coverage
Fixes:
- Add storage gap (uint256[49] __gap) to Shadow.sol for upgrade safety
- Change blockNumber type from uint48 to uint64 to match circuit's u64

Tests added:
- Fee calculation boundary cases (1 wei, 1000 wei, 1001 wei)
- Multiple claims with different nullifiers
- Upgrade scenario with storage preservation
- Old block verification (acceptable per PRD)
- Future block rejection

Updates CONTRACT_AUDIT.md to reflect all resolved issues.

+++ packages/contracts/src/iface/IShadow.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.33;
 
 interface IShadow {
     struct PublicInput {
-        uint48 blockNumber;
+        uint64 blockNumber;
         uint256 chainId;
         uint256 amount;
         address recipient;
